# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto controle de Filme- Senai
  description: |-
   API destinada ao controle de filmes para o projeto "bem legal"
  termsOfService: http://swagger.io/terms/
  contact:
    email: seunome@email.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.4.26
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://
paths:
  /v1/senai/locadora/filme:
    post:
      tags:
        - filme
      summary: Adiciona um novo Filme
      description: Endpoint para enviar os dados do filme a ser cadastrado
      requestBody:
        description: JSON contendo os dados do novo filme a ser inserido no sistema da locadora
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '415':
          description: ERROR content type
        '500':
         description: Error internal server(controller/model)
    get:
      tags:
        - filme
      summary: Lista todos os filmes
      description: Listagem de todos os filmes registrados no sistema da locadora
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'          
           
        '404':
          description: Not Found
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - filme
      summary: Busca um filme pelo ID
      description: Retorna um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca pelo Filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
        '500':
          description: ERROR internal server (model/controller)
    put:
      tags:
        - filme
      summary: atualiza um filme pelo ID
      description: buscar e atualizar um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca pelo ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: dados para atualizar um filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: not found  
        '415':
          description: ERROR content type
        '500':
          description: ERROR internal server (model/controller) 
    delete:
      tags:
        - filme
      summary: excluir um filme pelo id
      description: buscar e excluir um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
        '500':
          description: ERROR internal server (model/controller)
components:
  schemas:
    Filme:
      required:
        - nome
        - sinopse
        - capa
        - data_lancamento
        - duracao

      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sinopse:
          type: string
        capa:
          type: string
        data_lancamento:
          type: string
          format: date
        duracao:
          type: string
        valor:
          type: number
          format: double
        avaliacao:
          type: integer
  requestBodies:
    Filme:
      description: Objeto que deve ser enviado no Body das requisicoes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Filme'
    